enable ranking with prio service using promethee
