+++ org.planqk.nisq.analyzer.core/src/main/java/org/planqk/nisq/analyzer/core/prioritization/electre/ElectreIIIMethod.java
@@ -85,7 +85,7 @@ public class ElectreIIIMethod implements McdaMethod {
 
     @Override
     public String getDescription() {
-        return "TODO";
+        return "Using the MCDA method ELECTRE III to rank alternatives.";
     }
 
     @Override
